Update patch set 1
Patch Set 1:

Although this change is simpler, it is part of a series of patches adding AArch64 support to the remaining TA RPM packages (atm, that means ipa-deployer + all CAAS packages) [1].

It is a bit hard for me to add the proper reviewers to each of the change, so if you see changes in that topic touching one of the repos you contributed, please feel free to comment on them. Also, please ignore the spam review requests for the repos you are not familiar with.

[1] https://gerrit.akraino.org/r/q/topic:%2522aarch64-support%2522+status:open

Patch-set: 1
Reviewer: Gerrit User 137 <137@4cada01b-0314-4fe0-84a1-f9ddc89863f9>
Label: Verified=0